调整触摸事件逻辑，把滑动事件和点击事件放到base类中。

+++ app/src/main/cpp/jni/native-lib.cpp
@@ -34,7 +34,7 @@ Java_com_example_learnopengl_MyNativeRenderer_nativeProcessInput(JNIEnv *env, jo
 extern "C"
 JNIEXPORT void JNICALL
 Java_com_example_learnopengl_MyNativeRenderer_nativeMoveCallback(JNIEnv *env, jobject thiz,
-                                                                 jdouble x, jdouble y) {
+                                                                 jfloat x, jfloat y) {
     MyGLRender::GetInstance()->MoveCallback(x, y);
 }
 extern "C"
